core/types: moved to merkle.go, added comments, fixed issues
