+++ core/types/hashing.go
@@ -18,9 +18,6 @@ package types
 
 import (
 	"bytes"
-	"encoding/binary"
-	"fmt"
-	"math/bits"
 	"sync"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -113,117 +110,4 @@ func DeriveSha(list DerivableList, hasher TrieHasher) common.Hash {
 		hasher.Update(indexBuf, value)
 	}
 	return hasher.Hash()
-}
-
-func nextPowerOfTwo(v uint32) uint32 {
-	v--
-	v |= v >> 1
-	v |= v >> 2
-	v |= v >> 4
-	v |= v >> 8
-	v |= v >> 16
-	v++
-	return v
-}
-
-func MerkleizeAndMix(list DerivableList) common.Hash {
-	return mixinLength(Merkleize(list), list.Len())
-}
-
-func Merkleize(list DerivableList) common.Hash {
-	valueBuf := encodeBufferPool.Get().(*bytes.Buffer)
-	defer encodeBufferPool.Put(valueBuf)
-
-	sha := hasherPool.Get().(crypto.KeccakState)
-	defer hasherPool.Put(sha)
-
-	size := nextPowerOfTwo(uint32(list.Len()))
-	roots := make([]common.Hash, size)
-	for i := 0; i < list.Len(); i++ {
-		valueBuf.Reset()
-		list.EncodeIndex(i, valueBuf)
-		roots[i] = chunkAndHashSSZ(valueBuf.Bytes())
-	}
-	return merkleize(roots)
-}
-
-func merkleize(values []common.Hash) common.Hash {
-	// Check if len(values) is power of two
-	if len(values)&(len(values)-1) != 0 || len(values) == 0 {
-		panic(fmt.Sprintf("not a power of 2: %v", len(values)))
-	}
-	if len(values) == 1 {
-		return values[0]
-	}
-	round := make([]common.Hash, len(values))
-	copy(round, values)
-	for i := 0; i < bits.Len(uint(len(values)))-1; i++ {
-		for k := 0; k < len(round)/2; k++ {
-			round[k] = crypto.Keccak256Hash(round[k*2][:], round[k*2+1][:])
-		}
-		round = round[0 : len(round)/2]
-	}
-	return round[0]
-}
-
-func chunkAndHashSSZ(input []byte) common.Hash {
-	size := nextPowerOfTwo(uint32(len(input)))
-	roots := make([]common.Hash, size)
-	for start := 0; start < len(input); start += common.HashLength {
-		end := start + common.HashLength
-		if end > len(input) {
-			end = len(input)
-		}
-		copy(roots[start/common.HashLength][:], input[start:end])
-	}
-	return mixinLength(merkleize(roots), len(input))
-}
-
-func mixinLength(root common.Hash, length int) (h common.Hash) {
-	sha := hasherPool.Get().(crypto.KeccakState)
-	defer hasherPool.Put(sha)
-
-	var buf []byte
-	binary.LittleEndian.PutUint64(buf, uint64(length))
-
-	sha.Reset()
-	sha.Write(root[:])
-	sha.Write(buf[:])
-	sha.Read(h[:])
-	return h
-}
-
-var lookup map[int]common.Hash
-
-func init() {
-	// initialize the lookup
-	max := 2 ^ 30
-	base := [32]byte{}
-	prev := crypto.Keccak256Hash(base[:])
-	lookup[1] = prev
-	for i := 2; i < max; i = i << 1 {
-		prev = crypto.Keccak256Hash(prev.Bytes(), prev.Bytes())
-		lookup[i] = prev
-	}
-}
-
-func MerkleWithLimit(list DerivableList, limit int) (h common.Hash) {
-	if limit == 0 {
-		panic("should not happen")
-	}
-	if list.Len() <= limit/2 {
-		// TODO use merkleize, change to bytes based,
-		return Merkleize(list)
-	}
-	left := MerkleWithLimit(list, limit/2)
-	right := lookup[limit/2]
-
-	sha := hasherPool.Get().(crypto.KeccakState)
-	defer hasherPool.Put(sha)
-
-	sha.Reset()
-	sha.Write(left[:])
-	sha.Write(right[:])
-	sha.Read(h[:])
-	return h
 }
